perf(rust, python): improve list explode perf
