+++ polars/polars-core/src/chunked_array/list/mod.rs
@@ -5,6 +5,19 @@ use crate::chunked_array::Settings;
 use crate::prelude::*;
 
 impl ListChunked {
+    /// Get the inner data type of the list.
+    pub fn inner_dtype(&self) -> DataType {
+        match self.dtype() {
+            DataType::List(dt) => *dt.clone(),
+            _ => unreachable!(),
+        }
+    }
+
+    pub fn set_inner_dtype(&mut self, dtype: DataType) {
+        assert_eq!(dtype.to_physical(), self.inner_dtype().to_physical());
+        let field = Arc::make_mut(&mut self.field);
+        field.coerce(DataType::List(Box::new(dtype)));
+    }
     #[cfg(feature = "private")]
     pub fn set_fast_explode(&mut self) {
         self.bit_settings.insert(Settings::FAST_EXPLODE_LIST)
